Refactoring of k8s helpers
