+++ middleware/pkg/strings/slice.go
@@ -1,0 +1,11 @@
+package strings
+
+// StringInSlice check whether string a is a member of slice.
+func StringInSlice(a string, slice []string) bool {
+	for _, b := range slice {
+		if b == a {
+			return true
+		}
+	}
+	return false
+}
