+++ middleware/pkg/strings/slice_test.go
@@ -1,4 +1,4 @@
-package util
+package strings
 
 import (
 	"testing"
@@ -28,28 +28,6 @@ func TestStringInSlice(t *testing.T) {
 		actualResult := StringInSlice(example.String, example.Slice)
 		if actualResult != example.ExpectedResult {
 			t.Errorf("Expected stringInSlice result '%v' for example string='%v', slice='%v'. Instead got result '%v'.", example.ExpectedResult, example.String, example.Slice, actualResult)
-		}
-	}
-}
-
-// Test data for TestSymbolContainsWildcard cases.
-var testdataSymbolContainsWildcard = []struct {
-	Symbol         string
-	ExpectedResult bool
-}{
-	{"mynamespace", false},
-	{"*", true},
-	{"any", true},
-	{"my*space", true},
-	{"*space", true},
-	{"myname*", true},
-}
-
-func TestSymbolContainsWildcard(t *testing.T) {
-	for _, example := range testdataSymbolContainsWildcard {
-		actualResult := SymbolContainsWildcard(example.Symbol)
-		if actualResult != example.ExpectedResult {
-			t.Errorf("Expected SymbolContainsWildcard result '%v' for example string='%v'. Instead got result '%v'.", example.ExpectedResult, example.Symbol, actualResult)
 		}
 	}
 }
